ref: apply andr macros to fixed multi sig, boolean, form, string storage and cross chain swap
