+++ contracts/data-storage/andromeda-boolean/Cargo.toml
@@ -27,7 +27,6 @@ testing = ["cw-multi-test", "andromeda-testing"]
 cosmwasm-std = { workspace = true }
 cosmwasm-schema = { workspace = true }
 cw-storage-plus = { workspace = true }
-cw-utils = { workspace = true }
 cw20 = { workspace = true }
